get_api/requirements.txt
azure-core==1.32.0
azure-storage-blob==12.24.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
DateTime==5.5
idna==3.10
isodate==0.7.2
numpy==2.1.3
pandas==2.2.3
pathlib==1.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.1.1
